[FIX] sale: no warn onchange pricelist and no line
The popup warning that changing the pricelist will not update the existing order lines should not be diplayed it there is no order line set.

It was the case until 8.0. The change of behavior of on changes has broken this feature: Now, the on change return [(6, 0, [])] for an empy order line list, instead of just [].