[FIX] website_sale: make the 'Ship to a different address' always available
In the backend the option 'Allow a different address for delivery and invoicing' will set the group sale.group_delivery_invoice_address to employees. However the public user is not an employee (and we don't want to change that) so does not get this group. As we don't have similar group mechanism for public user, we have no other choice than always displaying the option to have different address. opw 610118