[FIX] web: wrong company logo in multi-company env
Don't preload company logo of the wrong company

When a user other than the admin signed in
and was in a company different than the admin
and this company had a different logo,
the company logo (in the upper left corner)
displayed the logo of the admin company,
instead of the logo of the company of this
user.

As soon as the user refreshed the page
in his browser, the issue was resolved,
but the logo should be the good one
from the beginning.

Closes #1453