[IMP] purchase: method action_picking_create return picking_id
Returning `picking_id` can be useful for overrides.

In addition, when there is no return statement,
the method basically returns None.
As this is a public method (not beginning with '_'), it can
be called with an xmlrpc call, and `None` is not an
accepted return value for the xmlrpc protocol.

Closes #1714