[FIX] point_of_sale: report, product_id of line is product_product
product_id column of pos_order_line is a product_product
the left join of l.product_id was done on product_template, instead of product_product

It worked as long as the ids product_product were the same as product_template. Meaning that, if you used variants, this report view was screwed.