[FIX] web: failing onchanges should not prevent to save
This patch is related to 82adba4714d9b809aad835d0d77e6b6f9fde8ddc

With the above patch, it wasn't possible anymore to save if an onchange failed. This isn't the expected behavior.
Besides, $.when.apply($, defs) is rejected as soon as one def fails, without waiting other defs to be either resolved or rejected.

This new patch allows to save if onchange fails, and wait for onchanges sequentially.